+++ tests/Functional/app/AppKernel.php
@@ -4,7 +4,6 @@ namespace Tests\Functional\app;
 
 use Psr\Log\NullLogger;
 use Symfony\Component\Config\Loader\LoaderInterface;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Kernel;
 
@@ -27,13 +26,5 @@ class AppKernel extends Kernel
     protected function build(ContainerBuilder $container): void
     {
         $container->register('logger', NullLogger::class);
-        $container->addCompilerPass(new class() implements CompilerPassInterface {
-            public function process(ContainerBuilder $container): void
-            {
-                // until https://github.com/doctrine/DoctrineBundle/pull/1263 is released on 1.12.x as well
-                $container->getDefinition('doctrine.dbal.logger.chain.default')->removeMethodCall('addLogger');
-                $container->getDefinition('doctrine.dbal.logger.chain')->removeMethodCall('addLogger');
-            }
-        });
     }
 }